renamed: navbar & footer
